+++ HardwareObjects/XMLRPCServer.py
@@ -155,7 +155,7 @@ class XMLRPCServer(HardwareObject):
         Method inherited from HardwareObject, called by framework-2.
         """
 
-        self.all_interfaces = self.getProperty('all_interfaces', "False")
+        self.all_interfaces = self.getProperty("all_interfaces", False)
         # Listen on all interfaces if <all_interfaces>True</all_interfaces>
         # otherwise only on the interface corresponding to socket.gethostname()
         if self.all_interfaces:
